xperience/scripts/requirements.txt
h5py==3.12.1
numpy==2.2.2
Pillow>=10.0.0
